egl/x11: ensure X creates pbuffers with non-zero size
The X Server visuals specify a maximum PBuffer width and height of zero.
When Mesa tries to allocate a zero size pixmap for the PBuffer, the request
is ignored, causing the subsequent DRI2 get buffers request to fail.
Patch the Mesa X11 EGL driver, so that when the pixmap backing the
PBuffer is created, 1 is substitued for a width or height of 0.

Signed-off-by: Randolph Sapp <rs@ti.com>