dri2: add a driver compatibility check extension
Add a DRI extension, DRI_DRIVER_COMPATIBILITY, to determine the
compatibilty of the render and display GPUs. The extension consists
of a single function, which is only called if the render and display
GPU FDs are different. If the function is not present, and the FDs
are different, the GPUs are assumed to be incompatible, as before
(i.e. intermediate linear buffers and blits may be needed to
transfer data between them).

Signed-off-by: Randolph Sapp <rs@ti.com>