loader/dri3: indicate to driver when drawable is a pixmap
When a new drawable is a pixmap, and the image driver supports
the createNewDrawableType entry point, indicate to the driver if
the drawable is a pixmap or not.

With GLX, pixmaps may be created using double buffered configs,
but rendering needs to be done to the front buffer. Providing a
hint that a drawable is actually a pixmap, makes it easier to
ensure this is done.

Signed-off-by: Randolph Sapp <rs@ti.com>